created mailbox folder for orginization
