--- src/main/java/frc/robot/commands/RunBelts.java
@@ -1,50 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-/*
- * Asimov's Laws:
- * The First Law: A robot may not injure a human being or, through inaction, allow a human being to come to harm.
- * The Second Law: A robot must obey the orders given it by human beings except where such orders would conflict with the First Law.
- * The Third Law: A robot must protect its own existence as long as such protection does not conflict with the First or Second Law.
- */
-
-package frc.robot.commands;
-
-import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.RobotContainer;
-import frc.robot.subsystems.mailbox.MailboxBelts;
-
-/** Command that activates belts when started, and deactivates belts when ended. */
-public class RunBelts extends Command {
-  private MailboxBelts mailboxBelts;
-
-  /** Creates a new DeployBelts. */
-  public RunBelts() {
-    this.mailboxBelts = RobotContainer.mailboxBelts;
-    addRequirements(mailboxBelts);
-    // Use addRequirements() here to declare subsystem dependencies.
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {
-    mailboxBelts.runBelts();
-  }
-
-  // Called every time the scheduler runs while the command is scheduled.
-  @Override
-  public void execute() {}
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {
-    mailboxBelts.stop();
-  }
-
-  // Returns true when the command should end.
-  @Override
-  public boolean isFinished() {
-    return false;
-  }
-}
