+++ src/main/java/frc/robot/commands/mailbox/RunBelts.java
@@ -9,26 +9,27 @@
  * The Third Law: A robot must protect its own existence as long as such protection does not conflict with the First or Second Law.
  */
 
-package frc.robot.commands;
+package frc.robot.commands.mailbox;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.RobotContainer;
-import frc.robot.subsystems.mailbox.MailboxPneumatics;
+import frc.robot.subsystems.mailbox.MailboxBelts;
 
-/** Command that extends the mailbox when started, and lowers the mailbox when ended. */
-public class DeployPneumatics extends Command {
-  private MailboxPneumatics mailboxPneumatics;
+/** Command that activates belts when started, and deactivates belts when ended. */
+public class RunBelts extends Command {
+  private MailboxBelts mailboxBelts;
 
-  /** Creates a new DeployPneumatics. */
-  public DeployPneumatics() {
-    this.mailboxPneumatics = RobotContainer.mailboxPneumatics;
-    addRequirements(mailboxPneumatics);
+  /** Creates a new DeployBelts. */
+  public RunBelts() {
+    this.mailboxBelts = RobotContainer.mailboxBelts;
+    addRequirements(mailboxBelts);
+    // Use addRequirements() here to declare subsystem dependencies.
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    mailboxPneumatics.extend();
+    mailboxBelts.runBelts();
   }
 
   // Called every time the scheduler runs while the command is scheduled.
@@ -38,6 +39,7 @@ public class DeployPneumatics extends Command {
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
+    mailboxBelts.stop();
   }
 
   // Returns true when the command should end.
